[RSC-33] Refactor transaction validating
This commit refactors `validateSum` (now called `validateTxPure`) so
that it now disallows zero/negative values in transactions, as well as
those that have empty inputs or outputs, or both.

It also adds a `canonizeTx` function that turns a transaction with the
previous invalid traits into a correct one if possible.