[RSC-224] Refactor indentity testing
This commit refactors identity testing in the following way:

1) 'Small' types have been added to reduce the test's times. In order
to verify that encoding/decoding a type to a certain type, @gromak
suggested it was not necessary to generate very large structures.
As such, wherever justifiable, 'Small' types are used.

2) All the identity tests have been moved to a separate folder since
they are all so similar, for ease of access.

3) Identity tests for types from third-party libraries have been deleted
since the libraries' authors implementations can be trusted.